requirements.txt
google-adk
fastapi
uvicorn[standard]
python-multipart
pymupdf
python-dotenv
ollama
pydantic
pydantic-settings
litellm
